Got Phong working almost correctly, need to check normal and Intersect

+++ RayTracer/Ray.h
@@ -19,7 +19,7 @@ public:
 
 	glm::vec3& getOrigin();
 	glm::vec3& getDirection();
-	glm::ivec3 castRay(const glm::vec3& _origin, const glm::vec3& _direction, std::vector<Sphere>& _spheres, const std::vector<Light>& _lights);
+	glm::vec3 castRay(const glm::vec3& _origin, const glm::vec3& _direction, std::vector<Sphere>& _spheres, const std::vector<Light>& _lights);
 	bool sceneIntersects(const glm::vec3& _origin, const glm::vec3& _direction, std::vector<Sphere>& _spheres, glm::vec3& _hit, glm::vec3& N, Material& _material) const;
 	glm::vec3 reflect(const glm::vec3& I, const glm::vec3& N);
 };
